Hackathon submission from Claude Wild (#36)
Team: Claude Wild

This change was the initial step in creating a more user-friendly method to access images from the Camera Roll. This change is currently for iOS and allows a user access to the last 30 images saved to their Camera Roll when the attachment button is selected. The user can scroll through the images and select one to be uploaded to the post.